==> 0_Download.ipynb <==
from radiant_mlhub import Collection
import tarfile
import os
from pathlib import Path
import json

os.environ['MLHUB_API_KEY'] = 'password'

collections = [
    #'ref_south_africa_crops_competition_v1_train_labels',
    #'ref_south_africa_crops_competition_v1_train_source_s1', # Comment this out if you do not wish to download the Sentinel-1 Data
    #'ref_south_africa_crops_competition_v1_train_source_s2',
    #'ref_south_africa_crops_competition_v1_test_labels',
    #'ref_south_africa_crops_competition_v1_test_source_s1', # Comment this out if you do not wish to download the Sentinel-1 Data
    'ref_south_africa_crops_competition_v1_test_source_s2'
]

def download(collection_id):
    print(f'Downloading {collection_id}...')
    collection = Collection.fetch(collection_id)
    path = collection.download('.')
    tar = tarfile.open(path, "r:gz")
    tar.extractall()
    tar.close()
    os.remove(path)
    
for c in collections:
    download(c)
