==> 1_Data preparation.ipynb <==
import datetime
import rasterio
import numpy as np
import pandas as pd
import geopandas as gpd

## 1. List of assets from Radiant MLHub


The Radiant MLHub API gives access to open Earth imagery training data for machine learning applications. You can learn more about the repository at the [Radiant MLHub site](https://mlhub.earth) and about the organization behind it at the [Radiant Earth Foundation site](https://radiant.earth).

Full documentation for the API is available at [docs.mlhub.earth](docs.mlhub.earth).

Each item in our collection is explained in json format compliant with [STAC](https://stacspec.org/) [label extension](https://github.com/radiantearth/stac-spec/tree/master/extensions/label) definition.

We create a DataFrame that sumarizes all the different assets.

import importlib
import create_asset_files as crass
importlib.reload(crass)

labels_path = {"train": 'ref_south_africa_crops_competition_v1_train_labels',
              "test": 'ref_south_africa_crops_competition_v1_test_labels'}

# creating asset collection for the train set
case = "train"
assets_train_df = crass.create_assets(labels_path[case])

assets_train_df.datetime = pd.to_datetime(assets_train_df.datetime)
assets_train_df = assets_train_df.assign(
    date = lambda x: x['datetime'].dt.date,
    month = lambda x: x['datetime'].dt.month.astype('Int64'),
    dayofyear = lambda x: x['datetime'].dt.dayofyear.astype('Int64'),
)

assets_train_df.tile_id = assets_train_df.tile_id.astype(int)
assets_train_df.to_csv(f"data/assets_{case}.csv", index = False)
print(f"assets_{case}.csv was saved in /data")
assets_train_df.head()

# creating asset collection for the test set
case = "test"
assets_test_df = create_assets(labels_path[case])

assets_test_df.datetime = pd.to_datetime(assets_test_df.datetime)
assets_test_df = assets_test_df.assign(
    date = lambda x: x['datetime'].dt.date,
    month = lambda x: x['datetime'].dt.month.astype('Int64'),
    dayofyear = lambda x: x['datetime'].dt.dayofyear.astype('Int64'),
)

assets_test_df.tile_id = assets_test_df.tile_id.astype(int)
assets_test_df.to_csv(f"data/assets_{case}.csv", index = False)
print(f"assets_{case}.csv was saved in /data")
assets_test_df.head()

## 2. Fields DataFrame

Now we want to generate a dataframe containing the inforatio of all the different fields.

assets_train_df = pd.read_csv('data/assets_train.csv')
fields_train = crass.create_field_list_train(assets_train_df)
fields_train.head()

len(fields_train["tile_id"].unique())

assets_test_df = pd.read_csv('data/assets_test.csv')
fields_test = crass.create_field_list_test(assets_test_df)
fields_test.head()

Some fields with 0 label and 0 field_id are present, they seem to be very small, so we will drop them from the list

fields_train = crass.remove_dupl_fields(fields_train)

fields_test = crass.remove_dupl_fields(fields_test)

We also add another column with the fields elevation over the sea in m.

crass.download_elevation(fields_train)

elev_src = rasterio.open('/home/jupyter/NF-Capstone-Crop-Classification/data/elev_merged.tif')
elev_read = elev_src.read()

We can save the DataFrame after that

fields_train['elevation'] = fields_train.geometry.apply(crass.get_elev, args=(elev_src, elev_read, ))
fields_train.to_file('data/fields_train.geojson', driver='GeoJSON')

fields_test['elevation'] = fields_test.geometry.apply(crass.get_elev, args=(elev_src, elev_read, ))
fields_test.to_file('data/fields_test.geojson', driver='GeoJSON')

## 3. Tile DataFrame

Now we want to extract all the information regarding the Tiles.

First, we check which days the images contain clouds.

fields_train = gpd.read_file('data/fields_train.geojson')
fields_test = gpd.read_file('data/fields_test.geojson')

assets_train_df = pd.read_csv('data/assets_train.csv')
assets_test_df = pd.read_csv('data/assets_test.csv')

import tile_utils as tilu
importlib.reload(tilu)

tiles_train = tilu.create_basic_tile_df(assets_train_df)
tiles_test = tilu.create_basic_tile_df(assets_test_df)

sunny_train = tilu.create_sunny_df(assets_train_df)
tiles_train = tiles_train.merge(sunny_train, how='inner', on='tile_id')

We have included:
+ the days that the sentinel 1 has taken a photo
+ the days that the sentinel 2 has taken a photo
+ the days that the sentinel 2 has taken a photo with no clouds
+ the rate between the days with no clouds
+ the days that the sentinel 2 has taken a photo with no clouds and with at least the 80% of the picture in good state.

We alse are interested in knowing the neighbor tiles

tiles_train_crs32634 = tiles_train.to_crs(32634)

# apply with a threshold of 4000m
tiles_train['tiles_closest'] = tiles_train_crs32634.apply(tilu.tiles_closest, axis='columns', args=(tiles_train_crs32634, 4000,))

tiles_train['tile_label_dist'] = tiles_train.apply(tilu.label_distribution, axis='columns', args=(fields_train,))

tiles_train.tiles_closest = tiles_train.tiles_closest.apply(eval)


tiles_train['close_label_dist'] = tiles_train.apply(tilu.close_tiles_label_dist, axis='columns', args=(tiles_train,))

tiles_train.tiles_closest = tiles_train.tiles_closest.astype(str)
tiles_train.to_file('data/tiles_train.geojson', driver='GeoJSON')

sunny_test = tilu.create_sunny_df(assets_test_df)
tiles_test = tiles_test.merge(sunny_test, how='inner', on='tile_id')

tiles_train = tilu.read_tile_geojson('data/tiles_train.geojson')
tiles_train_crs32634 = tiles_train.to_crs(32634)
tiles_test_crs32634 = tiles_test.to_crs(32634)

# apply with a threshold of 4000m
tiles_test['tiles_closest'] = tiles_test_crs32634.apply(tilu.tiles_closest, axis='columns', args=(tiles_train_crs32634, 4000,))

and the most common crops in them

tiles_test.head()

#tiles_test.tiles_closest = tiles_test.tiles_closest.apply(eval)
tiles_test['close_label_dist'] = tiles_test.apply(tilu.close_tiles_label_dist, axis='columns', args=(tiles_train,))

tiles_test.tiles_closest = tiles_test.tiles_closest.astype(str)
tiles_test.to_file('data/tiles_test.geojson', driver='GeoJSON')

## 4. Stack Sentinel 2 Channels

import stack_utils as stul
importlib.reload(stul)

assets_train_df = pd.read_csv('data/assets_train.csv')
assets_test_df = pd.read_csv('data/assets_test.csv')

tiles_train = tilu.read_tile_geojson('data/tiles_train.geojson')
tiles_test = tilu.read_tile_geojson('data/tiles_test.geojson')

num_days = 16

assets_s2_stacked_16days_train = stul.create_stacked(assets_train_df ,tiles_train, 's2', num_days)

assets_s2_stacked_16days_train.to_csv('data/assets_s2_stacked_16days_train.csv')

num_days = 16

assets_s1_stacked_16days_train = stul.create_stacked(assets_train_df ,tiles_train, 's1', num_days)

assets_s1_stacked_16days_train.to_csv('data/assets_s1_stacked_16days_train.csv')

num_days = 16

assets_s2_stacked_16days_test = stul.create_stacked(assets_test_df ,tiles_test, 's2', num_days)

assets_s2_stacked_16days_test.to_csv('data/assets_s2_stacked_16days_test.csv')

assets_s1_stacked_16days_test = stul.create_stacked(assets_test_df ,tiles_test, 's1', num_days)

assets_s1_stacked_16days_test.to_csv('data/assets_s1_stacked_16days_test.csv')

## 5. Crop Fields and create Mean + Var

### Train

import crop_utils as crul
importlib.reload(crul)

assets_train_df = pd.read_csv('data/assets_train.csv')
tiles_train= tilu.read_tile_geojson('data/tiles_train.geojson')
fields_train = gpd.read_file('data/fields_train.geojson')

assets_s2_stacked_16days_train = pd.read_csv('data/assets_s2_stacked_16days_train.csv', index_col = 0)
assets_s1_stacked_16days_train = pd.read_csv('data/assets_s1_stacked_16days_train.csv', index_col = 0)

train_mean_var_s2 = crul.create_mean_var(assets_train_df, assets_s2_stacked_16days_train, fields_train, tiles_train, "s2")
train_mean_var_s2.to_csv('data/mean_var_s2_16days_train.csv', index=False)

train_mean_var_s2.to_csv('data/mean_var_s2_16days_train.csv', index=False)

train_mean_var_s1 = crul.create_mean_var(assets_train_df, assets_s1_stacked_16days_train, fields_train, tiles_train, "s1")
train_mean_var_s1.to_csv('data/mean_var_s1_16days_train.csv', index=False)

train_mean_var_s2

### Test

assets_test_df = assets_test_df = pd.read_csv('data/assets_test.csv')
tiles_test = tilu.read_tile_geojson('data/tiles_test.geojson')
fields_test = gpd.read_file('data/fields_test.geojson')

assets_s2_stacked_16days_test = pd.read_csv('data/assets_s2_stacked_16days_test.csv', index_col = 0)
assets_s1_stacked_16days_test = pd.read_csv('data/assets_s1_stacked_16days_test.csv', index_col = 0)

test_mean_var_s2 = crul.create_mean_var(assets_test_df, assets_s2_stacked_16days_test, fields_test, tiles_test, "s2")
test_mean_var_s2.to_csv('data/mean_var_s2_16days_test.csv', index=False)

train_mean_var_s2.to_csv('data/mean_var_s2_16days_train.csv', index=False)

test_mean_var_s1 = crul.create_mean_var(assets_test_df, assets_s1_stacked_16days_test, fields_test, tiles_test, "s1")
test_mean_var_s1.to_csv('data/mean_var_s1_16days_test.csv', index=False)

train_mean_var_s1.to_csv('data/mean_var_s1_16days_train.csv', index=False)

test_mean_var = crul.create_mean_var(assets_test_df, assets_stacked_16days_test, fields_test, tiles_test)
test_mean_var.to_csv('data/mean_var_8days_test.csv', index=False)
